backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
python-dotenv==1.0.0
django-filter==21.1
djoser==2.1.0
Pillow==9.0.0
gunicorn==20.1.0
drf-extra-fields==3.0.3